Move the tests into test-amp-inabox.js
